Refactor to use same code
