+++ actors/miner/tests/miner_actor_test_bitfield_queue.rs
@@ -5,37 +5,12 @@ use fvm_ipld_bitfield::BitField;
 use fvm_ipld_bitfield::MaybeBitField;
 use fvm_ipld_blockstore::MemoryBlockstore;
 use fvm_shared::clock::{ChainEpoch, QuantSpec, NO_QUANTIZATION};
-use std::collections::HashMap;
 use std::iter::FromIterator;
 
 mod util;
 use util::*;
 
 const TEST_AMT_BITWIDTH: u32 = 3;
-
-#[derive(Default)]
-pub struct BQExpectation {
-    expected: HashMap<ChainEpoch, Vec<u64>>,
-}
-
-impl BQExpectation {
-    fn add(&mut self, epoch: ChainEpoch, values: Vec<u64>) -> &mut BQExpectation {
-        self.expected.entry(epoch).or_insert(values);
-        self
-    }
-
-    fn equals(&mut self, q: &mut BitFieldQueue<MemoryBlockstore>) {
-        assert_eq!(self.expected.len() as u64, q.amt.count());
-
-        q.amt
-            .for_each_mut(|epoch, bitfield| {
-                let values = &self.expected[&(epoch as i64)];
-                assert_bitfield_equals(bitfield, values);
-                Ok(())
-            })
-            .unwrap();
-    }
-}
 
 #[test]
 fn adds_values_to_empty_queue() {
@@ -48,8 +23,8 @@ fn adds_values_to_empty_queue() {
 
     queue.add_to_queue_values(epoch, values).unwrap();
 
-    let mut bq_expectation = BQExpectation::default();
-    bq_expectation.add(epoch, values.to_vec()).equals(&mut queue);
+    let bq_expectation = BitfieldQueueExpectation::default();
+    bq_expectation.add(epoch, &values).equals(&queue);
 }
 
 #[test]
@@ -58,13 +33,13 @@ fn adds_bitfield_to_empty_queue() {
     let rt = h.new_runtime();
     let mut queue = empty_bitfield_queue(&rt, TEST_AMT_BITWIDTH);
 
-    let values = BitField::try_from_bits(vec![1, 2, 3, 4]).unwrap();
+    let values = [1, 2, 3, 4];
     let epoch = ChainEpoch::from(42);
 
-    queue.add_to_queue(epoch, &values).unwrap();
+    queue.add_to_queue(epoch, &BitField::try_from_bits(values).unwrap()).unwrap();
 
-    let mut bq_expectation = BQExpectation::default();
-    bq_expectation.add(epoch, values.iter().collect()).equals(&mut queue);
+    let bq_expectation = BitfieldQueueExpectation::default();
+    bq_expectation.add(epoch, &values).equals(&queue);
 }
 
 #[test]
@@ -82,13 +57,13 @@ fn quantizes_added_epochs_according_to_quantization_spec() {
         queue.add_to_queue_values(val as i64, [val]).unwrap();
     }
 
-    let mut bq_expectation = BQExpectation::default();
+    let bq_expectation = BitfieldQueueExpectation::default();
     // expect values to only be set on quantization boundaries
     bq_expectation
-        .add(ChainEpoch::from(3), [0, 2, 3].to_vec())
-        .add(ChainEpoch::from(8), [4, 7, 8].to_vec())
-        .add(ChainEpoch::from(13), [9].to_vec())
-        .equals(&mut queue);
+        .add(ChainEpoch::from(3), &[0, 2, 3])
+        .add(ChainEpoch::from(8), &[4, 7, 8])
+        .add(ChainEpoch::from(13), &[9])
+        .equals(&queue);
 }
 
 #[test]
@@ -102,8 +77,8 @@ fn merges_values_within_same_epoch() {
     queue.add_to_queue_values(epoch, [1, 3].to_vec()).unwrap();
     queue.add_to_queue_values(epoch, [2, 4].to_vec()).unwrap();
 
-    let mut bq_expectation = BQExpectation::default();
-    bq_expectation.add(epoch, [1, 2, 3, 4].to_vec()).equals(&mut queue);
+    let bq_expectation = BitfieldQueueExpectation::default();
+    bq_expectation.add(epoch, &[1, 2, 3, 4]).equals(&queue);
 }
 
 #[test]
@@ -118,8 +93,8 @@ fn adds_values_to_different_epochs() {
     queue.add_to_queue_values(epoch1, [1, 3].to_vec()).unwrap();
     queue.add_to_queue_values(epoch2, [2, 4].to_vec()).unwrap();
 
-    let mut bq_expectation = BQExpectation::default();
-    bq_expectation.add(epoch1, [1, 3].to_vec()).add(epoch2, [2, 4].to_vec()).equals(&mut queue);
+    let bq_expectation = BitfieldQueueExpectation::default();
+    bq_expectation.add(epoch1, &[1, 3]).add(epoch2, &[2, 4]).equals(&queue);
 }
 
 #[test]
@@ -157,9 +132,9 @@ fn pop_until_does_nothing_if_until_parameter_before_first_value() {
     // modified is false
     assert!(!modified);
 
-    let mut bq_expectation = BQExpectation::default();
+    let bq_expectation = BitfieldQueueExpectation::default();
     // queue remains the same
-    bq_expectation.add(epoch1, [1, 3].to_vec()).add(epoch2, [2, 4].to_vec()).equals(&mut queue);
+    bq_expectation.add(epoch1, &[1, 3]).add(epoch2, &[2, 4]).equals(&queue);
 }
 
 #[test]
@@ -185,9 +160,9 @@ fn pop_until_removes_and_returns_entries_before_and_including_target_epoch() {
     // values from first two epochs are returned
     assert_bitfield_equals(&next, &[1, 3, 5]);
 
-    let mut bq_expectation = BQExpectation::default();
+    let bq_expectation = BitfieldQueueExpectation::default();
     // queue only contains remaining values
-    bq_expectation.add(epoch3, [6, 7, 8].to_vec()).add(epoch4, [2, 4].to_vec()).equals(&mut queue);
+    bq_expectation.add(epoch3, &[6, 7, 8]).add(epoch4, &[2, 4]).equals(&queue);
 
     // subsequent call to epoch less than next does nothing
     let (next, modified) = queue.pop_until(epoch3 - 1).unwrap();
@@ -196,9 +171,9 @@ fn pop_until_removes_and_returns_entries_before_and_including_target_epoch() {
     // no values are returned
     assert_bitfield_equals(&next, &[]);
 
-    let mut bq_expectation = BQExpectation::default();
+    let bq_expectation = BitfieldQueueExpectation::default();
     // queue only contains remaining values
-    bq_expectation.add(epoch3, [6, 7, 8].to_vec()).add(epoch4, [2, 4].to_vec()).equals(&mut queue);
+    bq_expectation.add(epoch3, &[6, 7, 8]).add(epoch4, &[2, 4]).equals(&queue);
 
     // popping the rest of the queue gets the rest of the values
     let (next, modified) = queue.pop_until(epoch4).unwrap();
@@ -207,9 +182,9 @@ fn pop_until_removes_and_returns_entries_before_and_including_target_epoch() {
     // rest of values are returned
     assert_bitfield_equals(&next, &[2, 4, 6, 7, 8]);
 
-    let mut bq_expectation = BQExpectation::default();
+    let bq_expectation = BitfieldQueueExpectation::default();
     // queue is now empty
-    bq_expectation.equals(&mut queue);
+    bq_expectation.equals(&queue);
 }
 
 #[test]
@@ -227,9 +202,9 @@ fn cuts_elements() {
     let to_cut = MaybeBitField::from_iter([2, 4, 5, 6]).unwrap();
     queue.cut(&to_cut).unwrap();
 
-    let mut bq_expectation = BQExpectation::default();
+    let bq_expectation = BitfieldQueueExpectation::default();
     // 3 shifts down to 2, 99 down to 95
-    bq_expectation.add(epoch1, [1, 2, 95].to_vec()).equals(&mut queue);
+    bq_expectation.add(epoch1, &[1, 2, 95]).equals(&queue);
 }
 
 #[test]
@@ -242,9 +217,9 @@ fn adds_empty_bitfield_to_queue() {
 
     queue.add_to_queue(epoch, &BitField::new()).unwrap();
 
-    let mut bq_expectation = BQExpectation::default();
+    let bq_expectation = BitfieldQueueExpectation::default();
     // ensures we don't add an empty entry
-    bq_expectation.equals(&mut queue);
+    bq_expectation.equals(&queue);
 }
 
 fn empty_bitfield_queue_with_quantizing(
